+++ HISTORY.md
@@ -5,6 +5,7 @@ http://jsoneditoronline.org
 
 ## not yet released, version 2.3.3
 
+- Added support for search (Ctrl+F) in the code editor Ace.
 - Fixed a positioning issue of the action menu when in bootstrap modal.
   (thanks tsash).